MAJ fonction modify, supprime bien l'image précédente seulement quand une nouvelle img est chargé dans la req

+++ controllers/sauce.js
@@ -45,10 +45,15 @@ exports.modifySauce = (req, res, next) => {
         res.status(400).json({ message: "Non autorisé" });
       } else {
         Sauce.findOne({ _id: req.params.id }).then((sauce) => {
-          //On déclare le nom de l'ancienne img de la sauce
-          const previousImg = sauce.imageUrl.split("/")[4];
-          // On utilise la précedente déclaration pour la supprimer du dossier img
-          fs.unlink(`images/${previousImg}`, () => {});
+          //On cherche a savoir si la req dispose d'une image
+          //Si c'est le cas on supprime la précédente
+          const oldImg = sauceObject.imageUrl;
+          if (oldImg) {
+            //On déclare le nom de l'ancienne img de la sauce
+            const previousImg = sauce.imageUrl.split("/")[4];
+            // On utilise la précedente déclaration pour la supprimer du dossier img
+            fs.unlink(`images/${previousImg}`, () => {});
+          }
         });
         Sauce.updateOne(
           { _id: req.params.id },
